% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_consumption.R
\name{basic_consumption}
\alias{basic_consumption}
\title{Calculate consumption}
\usage{
basic_consumption(df)
}
\arguments{
\item{df}{The datahub dataframe containing the index}
}
\value{
The consumption TimeSeries
}
\description{
This function calculates the consumption based on an index
}
