\name{opinumrtoolkit-package}
\alias{opinumrtoolkit-package}
\alias{opinumrtoolkit}
\docType{package}
\title{
\packageTitle{opinumrtoolkit}
}
\description{
\packageDescription{opinumrtoolkit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{opinumrtoolkit}
\packageIndices{opinumrtoolkit}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{opinumrtoolkit}

Maintainer: \packageMaintainer{opinumrtoolkit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
